switch from linear actuator to valve
